refactor(tui): extract prompt display helpers to pure module
